drivers: net: cpsw: fix compier warning with d-cache usage
Fix compier warning with d-cache api usage as there is a data type mismatch

+++ drivers/net/cpsw.c
@@ -842,7 +842,8 @@ static int cpsw_recv(struct eth_device *dev)
 	int len;
 
 	while (cpdma_process(priv, &priv->rx_chan, &buffer, &len) >= 0) {
-		invalidate_dcache_range(buffer, buffer + PKTSIZE_ALIGN);
+		invalidate_dcache_range((unsigned long)buffer,
+				(unsigned long)(buffer + PKTSIZE_ALIGN));
 		NetReceive(buffer, len);
 		cpdma_submit(priv, &priv->rx_chan, buffer, PKTSIZE);
 	}
